packages/edm-lib/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "Cython>=0.25", "numpy>=1.24.0"]
build-backend = "setuptools.build_meta"

[project]
name = "edm"
version = "0.1.0"
description = "EDM track analysis library"
readme = "README.md"
requires-python = ">=3.12"
authors = [
    {name = "EDM Project Contributors"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=1.24.0",
    "librosa>=0.10.0",
    "beat-this @ git+https://github.com/CPJKU/beat_this.git",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "transformers>=4.30.0",
    "cleanlab>=2.0.0",
    "tensorboard>=2.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "mutagen>=1.47.0",
    "structlog>=24.1.0",
    "msaf>=0.1.80",
    "pyyaml>=6.0.0",
    "dvc>=3.0.0",
    "jams>=0.3.5",
    "mir_eval>=0.7",
    "defusedxml>=0.7.1",
    "mlflow>=2.10.0",
]

[project.optional-dependencies]
annotation-ui = [
    "label-studio>=1.10.0",
    "label-studio-sdk>=0.0.32",
]

[tool.setuptools.packages.find]
where = ["src"]
